added check that deleted item is a directory

+++ app/api_utils.py
@@ -97,8 +97,11 @@ else:
 def perm_remove_directory(folder_to_delete):
     folder_to_delete_path = os.path.join(DELETION_LOCATION, folder_to_delete)
     if os.path.exists(folder_to_delete_path):
-        shutil.rmtree(folder_to_delete_path)
-        debug_logger.debug(f'Removed directory: {folder_to_delete}')
+        if os.path.isdir(folder_to_delete_path)
+            shutil.rmtree(folder_to_delete_path)
+            debug_logger.debug(f'Removed directory: {folder_to_delete}')
+        else:
+            debug_logger.debug(f'Following is not a directory: {folder_to_delete_path}')
     else:
         debug_logger.debug(f'Directory does not exist: {folder_to_delete}')
